fix CLI build on no_cgo systems (basic no_cgo support for motionplan package) (#4946)
