+++ .github/workflows/test.yml
@@ -158,7 +158,7 @@ jobs:
         with:
           cache-dependency-path: cli/test-requirements.txt
       - run: |
-          go run ./cli/viam/ module create --local-only --name schematest --public-namespace ns
+          go run -tags no_cgo ./cli/viam/ module create --local-only --name schematest --public-namespace ns
           cat meta.json | jq '.models = [{"api": "x:y:z", "model": "x:y:z"}]' > meta2.json
           python -m venv .venv
           source .venv/bin/activate
