+++ components/arm/arm.go
@@ -1,5 +1,3 @@
-//go:build !no_cgo
-
 // Package arm defines the arm that a robot uses to manipulate objects.
 // For more information, see the [arm component docs].
 //
