+++ components/arm/client.go
@@ -1,5 +1,3 @@
-//go:build !no_cgo
-
 // Package arm contains a gRPC based arm client.
 package arm
