+++ components/arm/server.go
@@ -1,5 +1,3 @@
-//go:build !no_cgo
-
 // Package arm contains a gRPC based arm service server.
 package arm
