+++ motionplan/motionPlanner.go
@@ -1,5 +1,3 @@
-//go:build !no_cgo
-
 // Package motionplan is a motion planning library.
 package motionplan
